add ability to print status tracker with option 6p did other stuff lolz

+++ Files/chkStatus.sh
@@ -115,11 +115,11 @@ printf "Please type or scan PC ID, or type (c) for search by customer name: "
 read PCID
 echo
 
-if [ "${PCID:0:2}" != "ID" ]; then
-	PCIDLoc=$(mktemp)
-	echo $PCID > $PCIDLoc
-	PCID=$(sed -e 's/^/ID/' $PCIDLoc)
-	rm $PCIDLoc
+if [ "${PCID,,}" != "c" ] && [ "${PCID:0:2}" != "ID" ]; then
+		PCIDLoc=$(mktemp)
+		echo $PCID > $PCIDLoc
+		PCID=$(sed -e 's/^/ID/' $PCIDLoc)
+		rm $PCIDLoc
 fi
 
 if [ "$PCID" == "c" ]; then
